........
  r17858 | bugman | 2012-10-16 14:14:20 +0200 (Tue, 16 Oct 2012) | 5 lines
  
  Python 3 fix for the generic_fns.structure.geometric.angles_regular() function.
  
  Integer divisions no longer produce integers.
........

+++ generic_fns/structure/geometric.py
@@ -57,7 +57,7 @@ def angles_regular(inc=None):
     # Generate the increment values of v.
     v = zeros(inc/2+1, float64)
     val = 1.0 / float(inc/2)
-    for i in range(inc/2+1):
+    for i in range(int(inc/2+1)):
         v[i] = float(i) * val
 
     # Generate the distribution of spherical angles theta.
